+++ packages/stencil-store/src/types.ts
@@ -33,13 +33,12 @@ export interface Setter<T> {
   <P extends keyof T>(propName: P & string, value: T[P]): void;
 }
 
-export interface CreateStoreReturn<T> {
+export interface ObservableMap<T> {
   state: T;
   get: Getter<T>;
   set: Setter<T>;
   reset: () => void;
   subscribe(subscription: StoreSubscription<T> | StoreSubscriptionObject<T>): void;
-  computed: any;
 }
 
 export interface ComputedReturn<T> {
